requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
sentence-transformers==5.1.2
faiss-cpu==1.12.0
numpy==2.3.4
torch==2.9.0
transformers==4.57.1
pydantic==2.12.4
sqlite3